requirements.txt
python-telegram-bot==13.15
openai==0.27.2
google-auth==2.15.0
google-auth-oauthlib==0.7.0
google-auth-httplib2==0.1.0
google-api-python-client==2.56.0
httpx==0.23.0
httpcore==0.15.0
python-dotenv==1.0.0
python-docx==0.8.11
urllib3==1.26.12